+++ src/classification/naive-bayes/index.js
@@ -7,8 +7,10 @@
  * [Reference]: https://en.wikipedia.org/wiki/Naive_Bayes_classifier
  */
 import GaussianNaiveBayes from './gaussian';
+import MultinomialNaiveBayes from './multinomial';
 
 export default GaussianNaiveBayes;
 export {
-  GaussianNaiveBayes as GaussianNaiveBayes
+  GaussianNaiveBayes,
+  MultinomialNaiveBayes
 };