+++ src/classification/naive-bayes/multinomial.js
@@ -1,0 +1,18 @@
+/**
+ * Talisman classification/naive-bayes/multinomial
+ * ================================================
+ *
+ * Implementation of the Multinomial Naive-Bayes classifier.
+ *
+ * [Reference]: https://en.wikipedia.org/wiki/Naive_Bayes_classifier
+ */
+import AbstractDiscreteNaiveBayes from './discrete';
+
+/**
+ * The Multinomial Naive Bayes classifier.
+ *
+ * @constructor
+ */
+export default class MultinomialNaiveBayes extends AbstractDiscreteNaiveBayes {
+
+}